+++ src/parser/check_bracket.c
@@ -30,7 +30,7 @@ int				check_close(char **str, int i, int j)
 	str[i - 1][ft_strlen(str[i - 1]) - 1] == '|' || str[i - 1][ft_strlen(str[i -
 	1]) - 1] == '>' || str[i - 1][ft_strlen(str[i - 1]) - 1] == '('))
 		return (-1);
-	if (j != ft_strlen(str[i]) - 1 && str[i][j + 1] != ';' && str[i][j + 1] !=
+	if (j != (int)ft_strlen(str[i]) - 1 && str[i][j + 1] != ';' && str[i][j + 1] !=
 	'|' && str[i][j + 1] != ')' && str[i][j + 1] != '$' && str[i][j + 1] < 42
 	&& str[i][j + 1] > 57)
 		return (-1);
